host/requirements.txt
imutils
numpy
opencv-python
python-decouple
python-telegram-bot>=20.0a1
matplotlib
tensorflow-gpu